@@ -7,12 +7,11 @@
     "build": "yarn feed-generate && yarn site-build",
     "feed-generate": "node --require esbuild-register src/feed/generate-feed-command.ts",
     "register-index": "node --require esbuild-register src/feed/register-index-command.ts",
-    "site-build": "eleventy",
-    "site-serve": "eleventy --serve",
-    "lint": "eslint ./src ./tests --max-warnings 0 --ext .ts --cache --format friendly",
-    "lint-fix": "yarn lint --fix",
-    "type-check": "yarn tsc --noEmit",
+    "site-build": "node --require esbuild-register ./node_modules/.bin/eleventy --config=eleventy.config.ts",
+    "site-serve": "node --require esbuild-register ./node_modules/.bin/eleventy --config=eleventy.config.ts --serve",
+    "lint": "eslint ./src ./tests --max-warnings 0 --cache --format friendly --fix && tsc --noEmit",
     "test": "vitest run",
+    "test-coverage": "vitest run --coverage",
     "renovate-validate": "renovate-config-validator",
     "preinstall": "typesync || :"
   },
@@ -37,46 +36,52 @@
   "devDependencies": {
     "@11ty/eleventy": "^2.0.1",
     "@11ty/eleventy-fetch": "^4.0.0",
-    "@11ty/eleventy-img": "^3.1.1",
+    "@11ty/eleventy-img": "^4.0.2",
+    "@eslint/js": "^9.1.1",
     "@supercharge/promise-pool": "^3.1.0",
-    "@types/11ty__eleventy-img": "^2.0.4",
+    "@types/11ty__eleventy-img": "^4.0.0",
     "@types/async-retry": "^1.4.3",
-    "@types/eslint": "^8.4.1",
+    "@types/clean-css": "^4.2.11",
+    "@types/eslint": "^9.0.0",
     "@types/eslint-config-prettier": "^6.11.2",
     "@types/eslint-plugin-prettier": "^3.1.0",
+    "@types/eslint__js": "~8.42.3",
     "@types/html-minifier-terser": "^7.0.2",
     "@types/http-server": "^0.12.1",
     "@types/node": "^20.10.4",
     "@types/open-graph-scraper": "^4.8.1",
     "@types/prettier": "^3.0.0",
     "@types/request": "^2.48.8",
-    "@typescript-eslint/eslint-plugin": "^6.8.0",
-    "@typescript-eslint/parser": "^6.8.0",
+    "@typescript-eslint/eslint-plugin": "^8.0.0",
+    "@typescript-eslint/parser": "^8.0.0",
+    "@vitest/coverage-v8": "^2.0.3",
     "async-retry": "^1.3.3",
     "await-to-js": "^3.0.0",
     "axios": "^1.6.0",
+    "clean-css": "^5.3.3",
     "dayjs": "^1.11.10",
     "esbuild-register": "^3.5.0",
-    "eslint": "^8.52.0",
+    "eslint": "^9.1.1",
     "eslint-config-prettier": "^9.0.0",
     "eslint-formatter-friendly": "^7.0.0",
     "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-prettier": "^5.0.1",
     "fast-xml-parser": "^4.3.2",
     "feed": "^4.2.2",
-    "googleapis": "^129.0.0",
+    "googleapis": "^142.0.0",
     "html-minifier-terser": "^7.2.0",
     "http-server": "^14.1.1",
     "libxmljs": "^1.0.11",
     "log4js": "^6.9.1",
-    "open-graph-scraper": "^4.11.0",
+    "open-graph-scraper": "^6.5.1",
     "prettier": "^3.0.3",
     "renovate": "^37.253.1",
     "request": "^2.88.2",
     "rss-parser": "^3.12.0",
+    "sharp-ico": "^0.1.5",
     "typescript": "^5.2.2",
-    "typesync": "^0.12.0",
-    "vitest": "^1.0.4"
+    "typesync": "^0.13.0",
+    "vitest": "^2.0.0"
   },
   "prettier": {
     "tabWidth": 2,
@@ -91,5 +96,5 @@
     "node": ">=20",
     "yarn": ">=3.7.0"
   },
-  "packageManager": "yarn@3.8.1"
+  "packageManager": "yarn@4.4.0"
 }
