@@ -1,7 +1,7 @@
-const log4js = require('log4js');
+import * as log4js from 'log4js';
 
 const logger = log4js.getLogger();
 
-logger.level = 'all';
+logger.level = log4js.levels.ALL;
 
 export { logger };
